+++ SynthEBD/Classes_Core/ViewModels/BodyGenConfig SubModels/VM_BodyGenTemplateMenu.cs
@@ -110,7 +110,7 @@ namespace SynthEBD
         public bool bAllowUnique { get; set; }
         public bool bAllowNonUnique { get; set; }
         public bool bAllowRandom { get; set; }
-        public int ProbabilityWeighting { get; set; }
+        public double ProbabilityWeighting { get; set; }
         public ObservableCollection<VM_CollectionMemberString> RequiredTemplates { get; set; }
         public NPCWeightRange WeightRange { get; set; }
